refactor(scrollbar): reduce duplication and align pixel-aware drag
Extract shared ContentAwareScrollbarShell composable (-357 net lines)
hosting drag state, hover, theming, and gesture wiring. Both scrollbars
now own only their unique logic: ContentScrollbar keeps far-drag throttle
for paged windows; CommentariesScrollbar now uses pixel-aware drag matching
the shell's geometry (50% drag = 50% visual content, not 50% items).

- ScrollbarMetrics: add visualLinesOf, buildCumulativePixels, findLineIndexForPixel
- ContentAwareScrollbarShell: new reusable shell with convergence/timeout effects
- ContentScrollbar: 353→198 lines, delegates layout to shell
- CommentariesScrollbar: 268→139 lines, delegates layout to shell, adds pixel-aware drag
